code/backend/requirements.txt
# Sentinel-AI Backend
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai==1.50.0
numpy==1.26.4
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
httpx==0.27.0
sqlalchemy[asyncio]==2.0.32
asyncpg==0.29.0
aiosqlite==0.20.0
faiss-cpu==1.8.0
pytest==8.3.0
pytest-asyncio==0.24.0
